+++ dart/lib/vm_service_lib.dart
@@ -370,7 +370,8 @@ abstract class VmServiceInterface {
   /// `limit` is the maximum number of instances to be returned.
   ///
   /// See [InstanceSet].
-  Future<InstanceSet> getInstances(String objectId, int limit);
+  Future<InstanceSet> getInstances(
+      String isolateId, String objectId, int limit);
 
   /// The `getIsolate` RPC is used to lookup an `Isolate` object by its `id`.
   ///
@@ -508,7 +509,7 @@ abstract class VmServiceInterface {
   /// If `getVMTimeline` is invoked while the current recorder is one of Fuchsia
   /// or Systrace, the `114` error code, invalid timeline request, will be
   /// returned as timeline events are handled by the OS in these modes.
-  Future<Timeline> getVMTimeline(int timeOriginMicros, int timeExtentMicros);
+  Future<Timeline> getVMTimeline({int timeOriginMicros, int timeExtentMicros});
 
   /// The `getVMTimelineFlags` RPC returns information about the current VM
   /// timeline configuration.
@@ -842,6 +843,7 @@ class VmServerConnection {
           break;
         case 'getInstances':
           response = await _serviceImplementation.getInstances(
+            params['isolateId'],
             params['objectId'],
             params['limit'],
           );
@@ -892,8 +894,8 @@ class VmServerConnection {
           break;
         case 'getVMTimeline':
           response = await _serviceImplementation.getVMTimeline(
-            params['timeOriginMicros'],
-            params['timeExtentMicros'],
+            timeOriginMicros: params['timeOriginMicros'],
+            timeExtentMicros: params['timeExtentMicros'],
           );
           break;
         case 'getVMTimelineFlags':
@@ -1213,10 +1215,10 @@ class VmService implements VmServiceInterface {
   Future<AllocationProfile> getAllocationProfile(String isolateId,
       {bool reset, bool gc}) {
     Map m = {'isolateId': isolateId};
-    if (reset != null) {
+    if (reset != null && reset) {
       m['reset'] = reset;
     }
-    if (gc != null) {
+    if (gc != null && gc) {
       m['gc'] = gc;
     }
     return _call('getAllocationProfile', m);
@@ -1226,8 +1228,10 @@ class VmService implements VmServiceInterface {
   Future<FlagList> getFlagList() => _call('getFlagList');
 
   @override
-  Future<InstanceSet> getInstances(String objectId, int limit) {
-    return _call('getInstances', {'objectId': objectId, 'limit': limit});
+  Future<InstanceSet> getInstances(
+      String isolateId, String objectId, int limit) {
+    return _call('getInstances',
+        {'isolateId': isolateId, 'objectId': objectId, 'limit': limit});
   }
 
   @override
@@ -1300,11 +1304,15 @@ class VmService implements VmServiceInterface {
   Future<VM> getVM() => _call('getVM');
 
   @override
-  Future<Timeline> getVMTimeline(int timeOriginMicros, int timeExtentMicros) {
-    return _call('getVMTimeline', {
-      'timeOriginMicros': timeOriginMicros,
-      'timeExtentMicros': timeExtentMicros
-    });
+  Future<Timeline> getVMTimeline({int timeOriginMicros, int timeExtentMicros}) {
+    Map m = {};
+    if (timeOriginMicros != null) {
+      m['timeOriginMicros'] = timeOriginMicros;
+    }
+    if (timeExtentMicros != null) {
+      m['timeExtentMicros'] = timeExtentMicros;
+    }
+    return _call('getVMTimeline', m);
   }
 
   @override
@@ -1985,8 +1993,12 @@ class AllocationProfile extends Response {
   AllocationProfile._fromJson(Map<String, dynamic> json)
       : super._fromJson(json) {
     memoryUsage = createServiceObject(json['memoryUsage']);
-    dateLastAccumulatorReset = json['dateLastAccumulatorReset'];
-    dateLastServiceGC = json['dateLastServiceGC'];
+    dateLastAccumulatorReset = json['dateLastAccumulatorReset'] is String
+        ? int.parse(json['dateLastAccumulatorReset'])
+        : json['dateLastAccumulatorReset'];
+    dateLastServiceGC = json['dateLastServiceGC'] is String
+        ? int.parse(json['dateLastServiceGC'])
+        : json['dateLastServiceGC'];
     members =
         new List<ClassHeapStats>.from(createServiceObject(json['members']));
   }
@@ -2336,8 +2348,8 @@ class ClassHeapStats extends Response {
     instancesAccumulated = json['instancesAccumulated'];
     instancesCurrent = json['instancesCurrent'];
     classRef = createServiceObject(json['class']);
-    new_ = new List<int>.from(json['new']);
-    old = new List<int>.from(json['old']);
+    new_ = json['new'] == null ? null : new List<int>.from(json['new']);
+    old = json['old'] == null ? null : new List<int>.from(json['old']);
     promotedBytes = json['promotedBytes'];
     promotedInstances = json['promotedInstances'];
   }
@@ -3849,14 +3861,14 @@ class InstanceSet extends Response {
   int totalCount;
 
   /// An array of instances of the requested type.
-  List<InstanceRef> instances;
+  List<ObjRef> instances;
 
   InstanceSet();
 
   InstanceSet._fromJson(Map<String, dynamic> json) : super._fromJson(json) {
     totalCount = json['totalCount'];
-    instances =
-        new List<InstanceRef>.from(createServiceObject(json['instances']));
+    instances = new List<ObjRef>.from(
+        createServiceObject(json['instances'] ?? json['samples']));
   }
 
   @override