+++ dart/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
+## 3.20.0+2
+- allow optional params in `getVMTimeline`
+
 ## 3.20.0+1
 - handle null isolate ids in `callServiceExtension`
+- add backwards compatibility for `InstanceSet` and `AllocationProfile`
 
 ## 3.20.0
 - rev to 3.20.0; expose public methods added in 3.17 - 3.20 VM Service Protocol versions